feat: Enhance kernel panic handling with 64-bit support and detailed register dump feat: Add string concatenation function to string library refactor: Update memory management functions to support 64-bit operations

+++ src/kernel/panic.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 // Panic with detailed error information
-void kernel_panic(const char* message, uint32_t error_code);
+void kernel_panic(const char* message, uint64_t error_code);
 
 // Exception names
 extern const char* exception_messages[32];
